Change Long to integer of id param and getOrdersByUserId of OrderDao done

+++ src/main/java/fr/esgi/ticketapi/infrastructure/dataprovider/dao/MySQLOrderDao.java
@@ -27,12 +27,13 @@ public class MySQLOrderDao implements OrderDao {
     }
 
     @Override
-    public List<Order> getOrdersByUserId(Long userId) {
+    public List<Order> getOrdersByUserId(Integer userId) {
         return null;
     }
 
     @Override
-    public Order getOrderByOrderIdAndUserId(Long userId, Long orderId) {
+    public Order getOrderByOrderIdAndUserId(Integer userId, Integer orderId) {
         return null;
     }
+
 }
